backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
SQLAlchemy>=2.0.30
alembic>=1.13.0
psycopg[binary]>=3.1.12
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.0.3
PyJWT>=2.8.0
python-dotenv>=1.0.0
greenlet>=3.0.3
